moved NewPool out of internal package
