+++ drivers/pg/pg.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
+	"time"
 
 	"github.com/jackc/pgx/v5"
+	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/specterops/dawgs"
 	"github.com/specterops/dawgs/cypher/models/pgsql"
 	"github.com/specterops/dawgs/graph"
@@ -16,7 +18,8 @@ const (
 
 	// defaultBatchWriteSize is currently set to 2k. This is meant to strike a balance between the cost of thousands
 	// of round-trips against the cost of locking tables for too long.
-	defaultBatchWriteSize = 2_000
+	defaultBatchWriteSize     = 2_000
+	poolInitConnectionTimeout = time.Second * 10
 )
 
 func AfterPooledConnectionEstablished(ctx context.Context, conn *pgx.Conn) error {
@@ -44,6 +47,33 @@ func AfterPooledConnectionRelease(conn *pgx.Conn) bool {
 	}
 
 	return true
+}
+
+func NewPool(connString string) (*pgxpool.Pool, error) {
+	poolCtx, done := context.WithTimeout(context.Background(), poolInitConnectionTimeout)
+	defer done()
+
+	poolCfg, err := pgxpool.ParseConfig(connString)
+	if err != nil {
+		return nil, err
+	}
+
+	// TODO: Min and Max connections for the pool should be configurable
+	poolCfg.MinConns = 5
+	poolCfg.MaxConns = 50
+
+	// Bind functions to the AfterConnect and AfterRelease hooks to ensure that composite type registration occurs.
+	// Without composite type registration, the pgx connection type will not be able to marshal PG OIDs to their
+	// respective Golang structs.
+	poolCfg.AfterConnect = AfterPooledConnectionEstablished
+	poolCfg.AfterRelease = AfterPooledConnectionRelease
+
+	pool, err := pgxpool.NewWithConfig(poolCtx, poolCfg)
+	if err != nil {
+		return nil, err
+	}
+
+	return pool, nil
 }
 
 func init() {
